+++ server/src/resources/editor_resources.ts
@@ -1,10 +1,9 @@
-import { Resource } from 'fastmcp';
 import { getGodotConnection } from '../utils/godot_connection.js';
 
 /**
  * Resource that provides information about the current state of the Godot editor
  */
-export const editorStateResource: Resource = {
+export const editorStateResource = {
   uri: 'godot/editor/state',
   name: 'Godot Editor State',
   mimeType: 'application/json',
@@ -28,7 +27,7 @@ export const editorStateResource: Resource = {
 /**
  * Resource that provides information about the currently selected node
  */
-export const selectedNodeResource: Resource = {
+export const selectedNodeResource = {
   uri: 'godot/editor/selected_node',
   name: 'Godot Selected Node',
   mimeType: 'application/json',
@@ -52,7 +51,7 @@ export const selectedNodeResource: Resource = {
 /**
  * Resource that provides information about the currently edited script
  */
-export const currentScriptResource: Resource = {
+export const currentScriptResource = {
   uri: 'godot/editor/current_script',
   name: 'Current Script in Editor',
   mimeType: 'text/plain',