+++ server/src/resources/project_resources.ts
@@ -1,10 +1,9 @@
-import { Resource } from 'fastmcp';
 import { getGodotConnection } from '../utils/godot_connection.js';
 
 /**
  * Resource that provides information about the Godot project structure
  */
-export const projectStructureResource: Resource = {
+export const projectStructureResource = {
   uri: 'godot/project/structure',
   name: 'Godot Project Structure',
   mimeType: 'application/json',
@@ -28,7 +27,7 @@ export const projectStructureResource: Resource = {
 /**
  * Resource that provides project settings
  */
-export const projectSettingsResource: Resource = {
+export const projectSettingsResource = {
   uri: 'godot/project/settings',
   name: 'Godot Project Settings',
   mimeType: 'application/json',
@@ -52,7 +51,7 @@ export const projectSettingsResource: Resource = {
 /**
  * Resource that provides a list of all project resources
  */
-export const projectResourcesResource: Resource = {
+export const projectResourcesResource = {
   uri: 'godot/project/resources',
   name: 'Godot Project Resources',
   mimeType: 'application/json',